+++ include/upcie/dmabuf.h
@@ -1,35 +1,21 @@
 // SPDX-License-Identifier: BSD-3-Clause
 
 /**
- * Get physical addresses from a dma-buf
- * =====================================
+ * Representation of a dma-buf and its physical pages
+ * ==================================================
  *
- * This is a generic interface that is compatible with any dma-buf.
- * A dma-buf FD can be obtained either from host memory, memfd->udmabuf, or
- * device memory, e.g., CUDA or ROCm.
+ * A generic interface compatible with any dma-buf. A dma-buf descriptor can be
+ * obtained from host memory (memfd via udmabuf) or from device memory, e.g.
+ * CUDA or ROCm.
  *
- * NOTE: The import path uses the out-of-tree dmabuf_import module, which serves
- * its ioctls on /dev/dmabuf_import. Its UAPI is <linux/dmabuf_import.h>,
- * installed by the dmabuf-import DKMS package, which ships with the upcie
- * release as an asset but is versioned independently. That header is optional:
- * when it is not available, dmabuf_attach()/dmabuf_detach() compile as stubs
- * returning -ENOTSUP, so upcie builds and runs without the module (the import
- * calls simply fail). Install the package to enable the path; the module must
- * also be loaded at runtime.
+ * This header is dependency-free: it describes a dma-buf and segments its
+ * pages, and needs nothing beyond libc. Resolving the DMA addresses behind a
+ * dma-buf in the first place needs the out-of-tree dmabuf_import module and
+ * lives in <upcie/experimental/dmabuf_import.h>.
  *
  * @file dmabuf.h
  * @version 0.5.2
  */
-
-#include <linux/dma-buf.h>
-
-/* Optional: pull in the dmabuf_import UAPI if the DKMS package installed it.
- * Guarded with __has_include so upcie builds without it. */
-#if defined(__has_include)
-#  if __has_include(<linux/dmabuf_import.h>)
-#    include <linux/dmabuf_import.h>
-#  endif
-#endif
 
 struct dmabuf_page {
 	uint64_t addr;			///< Address of a page
@@ -102,125 +88,3 @@ dmabuf_get_lut(struct dmabuf *dmabuf, size_t nphys, uint64_t *phys_lut, uint64_t
 
 	return 0;
 }
-
-#ifdef DMABUF_IMPORT_ATTACH
-/**
- * Attach to dma-buf with given FD
- *
- * Populates the given dma-buf structure with information about the dma-buf.
- */
-static inline int
-dmabuf_attach(int dmabuf_fd, struct dmabuf *dmabuf)
-{
-	struct dmabuf_import_attach attach;
-	struct dmabuf_import_get_map *map = NULL;
-	int import_fd, err;
-	size_t map_size, pages_size;
-
-	import_fd = open(DMABUF_IMPORT_DEVPATH, O_RDWR);
-	if (import_fd < 0) {
-		err = -errno;
-		UPCIE_DEBUG("FAILED: open(%s), errno: %d", DMABUF_IMPORT_DEVPATH, err);
-		return err;
-	}
-
-	memset(&attach, 0, sizeof(attach));
-	attach.fd = dmabuf_fd;
-
-	err = ioctl(import_fd, DMABUF_IMPORT_ATTACH, &attach);
-	if (err) {
-		err = -errno;
-		UPCIE_DEBUG("FAILED: ioctl(DMABUF_IMPORT_ATTACH), errno: %d", err);
-		goto exit;
-	}
-
-	map_size = attach.count * sizeof(struct dmabuf_import_dma_map);
-	map = malloc(sizeof(struct dmabuf_import_get_map) + map_size);
-	if (!map) {
-		err = -errno;
-		UPCIE_DEBUG("FAILED: malloc(map), errno: %d", err);
-		ioctl(import_fd, DMABUF_IMPORT_DETACH, &dmabuf_fd);
-		goto exit;
-	}
-
-	memset(map, 0, sizeof(*map));
-	map->fd = dmabuf_fd;
-	map->count = attach.count;
-
-	err = ioctl(import_fd, DMABUF_IMPORT_GET_MAP, map);
-	if (err) {
-		err = -errno;
-		UPCIE_DEBUG("FAILED: ioctl(DMABUF_IMPORT_GET_MAP), errno: %d\n", err);
-		ioctl(import_fd, DMABUF_IMPORT_DETACH, &dmabuf_fd);
-		goto exit;
-	}
-
-	memset(dmabuf, 0, sizeof(*dmabuf));
-	dmabuf->fd = dmabuf_fd;
-	dmabuf->npages = map->count;
-	pages_size = sizeof(struct dmabuf_page) * dmabuf->npages;
-	dmabuf->pages = malloc(pages_size);
-	if (!dmabuf->pages) {
-		err = -errno;
-		UPCIE_DEBUG("FAILED: malloc(dmabuf->pages), errno: %d", err);
-		ioctl(import_fd, DMABUF_IMPORT_DETACH, &dmabuf_fd);
-		goto exit;
-	}
-
-	memcpy(dmabuf->pages, map->dma_arr, pages_size);
-
-exit:
-	free(map);
-	close(import_fd);
-	return err;
-}
-
-/**
- * Detach from given dma-buf
- *
- * NOTE: This doesn't free the underlying memory
- */
-static inline int
-dmabuf_detach(struct dmabuf *dmabuf)
-{
-	int import_fd, err;
-
-	free(dmabuf->pages);
-
-	import_fd = open(DMABUF_IMPORT_DEVPATH, O_RDWR);
-	if (import_fd < 0) {
-		err = -errno;
-		UPCIE_DEBUG("FAILED: open(%s), errno: %d", DMABUF_IMPORT_DEVPATH, err);
-		return err;
-	}
-
-	err = ioctl(import_fd, DMABUF_IMPORT_DETACH, &dmabuf->fd);
-	if (err) {
-		err = -errno;
-		UPCIE_DEBUG("FAILED: ioctl(DMABUF_IMPORT_DETACH), errno: %d\n", err);
-		// fall-through
-	}
-
-	close(dmabuf->fd);
-	close(import_fd);
-	return err;
-}
-#else /* !DMABUF_IMPORT_ATTACH: dmabuf-import UAPI unavailable, provide stubs */
-/**
- * Attach stub -- the dmabuf_import UAPI (<linux/dmabuf_import.h>) is not
- * available. Install the dmabuf-import DKMS package to enable importing.
- */
-static inline int
-dmabuf_attach(int UPCIE_UNUSED(dmabuf_fd), struct dmabuf *UPCIE_UNUSED(dmabuf))
-{
-	UPCIE_DEBUG("FAILED: dmabuf_import unavailable; install dmabuf-import-dkms");
-	return -ENOTSUP;
-}
-
-static inline int
-dmabuf_detach(struct dmabuf *UPCIE_UNUSED(dmabuf))
-{
-	UPCIE_DEBUG("FAILED: dmabuf_import unavailable; install dmabuf-import-dkms");
-	return -ENOTSUP;
-}
-#endif /* DMABUF_IMPORT_ATTACH */
