+++ include/upcie/hipmem_heap.h
@@ -6,6 +6,10 @@
  *
  * This heap implementation uses the HIP/ROCm runtime to pre-allocate memory and
  * the dma-buf interface to get the physical addresses.
+ *
+ * EXPERIMENTAL dependency: the physical addresses come from the out-of-tree
+ * dmabuf-import DKMS module, see <upcie/experimental/dmabuf_import.h>.
+ *
  * While the heap memory is allocated on the GPU, the freelist is maintained in
  * host memory. This avoids segfaults at the cost of slower alloc/free.
  *
@@ -107,7 +111,7 @@ hipmem_heap_term(struct hipmem_heap *heap)
 		return;
 	}
 
-	dmabuf_detach(&heap->dmabuf);
+	dmabuf_import_detach(&heap->dmabuf);
 	hipmem_heap_empty_freelist(heap->freelist);
 	free(heap->phys_lut);
 	hipFree((void *)heap->vaddr);
@@ -175,9 +179,10 @@ hipmem_heap_init(struct hipmem_heap *heap, size_t size, struct hipmem_config *co
 	heap->freelist->free = 1;
 	heap->freelist->next = NULL;
 
-	err = dmabuf_attach(dmabuf_fd, &heap->dmabuf);
+	/* NOTE: EXPERIMENTAL dependency, see <upcie/experimental/dmabuf_import.h> */
+	err = dmabuf_import_attach(dmabuf_fd, &heap->dmabuf);
 	if (err) {
-		UPCIE_DEBUG("FAILED: dmabuf_attach(), err: %d", err);
+		UPCIE_DEBUG("FAILED: dmabuf_import_attach(), err: %d", err);
 		goto error;
 	}
 
@@ -199,7 +204,7 @@ hipmem_heap_init(struct hipmem_heap *heap, size_t size, struct hipmem_config *co
 	return 0;
 
 error_after_attach:
-	dmabuf_detach(&heap->dmabuf);
+	dmabuf_import_detach(&heap->dmabuf);
 	free(heap->freelist);
 	free(heap->phys_lut);
 	hipFree(vaddr);
