+++ tests/test_hipmem_producer.c
@@ -8,9 +8,9 @@
  * Exercises the AMD/HIP half of the hipmem backend without the dma-buf import
  * path: query the device config (HIP + BAR1) and allocate GPU memory, then
  * export it as a P2P dma-buf via HSA. This is the producer that hipmem_heap
- * feeds into dmabuf_attach()/dmabuf_get_lut(); the full heap additionally needs
- * the dmabuf-import kernel module, which this check deliberately avoids so it
- * runs on a stock ROCm install.
+ * feeds into dmabuf_import_attach()/dmabuf_get_lut(); the full heap
+ * additionally needs the dmabuf-import kernel module, which this check
+ * deliberately avoids so it runs on a stock ROCm install.
  */
 #include <upcie/upcie_hip.h>
